Assembly debug comments, div/mul fix, EXPR2 optimization, jump op fix, start functions for Linux/OS X.
Added assembly debug comments to make it easier to find bugs.
These messages tend to make the assembly look messy, so TODO: debug levels perhaps?

Replaced the "mul" and "div" instructions (for floating point operations) with "imul"
and "idiv". rdx is also zeroed before using div (div operates on the 128bit "register" RDX:RAX).

Added "EXPR2" as optimization, which relies on the fact the first expression doesn't have an operator.
It skips some unnecessary assignments to temporary registers.

The jump ops are now "reversed": before, the branch after an if statement would be skipped if the
condition were true when it should only be skipped when the condition is false. "Reversing" the jump
ops fixes this.

Added code to generate separate start functions for both Linux and OS X.